requirements.txt
numpy>=1.24.0
pandas>=2.0.0
networkx>=3.0
matplotlib>=3.7.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
langchain>=0.0.267
pypdf>=3.15.0
python-dotenv>=1.0.0
scikit-learn>=1.2.0
plotly>=5.15.0
spacy>=3.6.0
